Move components into folders for more intuitive organization and update import paths.
