+++ app/frontend/components/Enrollment/Common/LinkRendering.jsx
