+++ app/frontend/components/Enrollment/Common/NewStudentError.jsx
