+++ app/frontend/components/Enrollment/Common/StudentText.jsx
