+++ app/frontend/components/Enrollment/ListLocationEnrollments.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import EndTime from '~/components/EndTime.jsx';
+import EndTime from '~/components/Enrollment/Common/EndTime.jsx';
 import { Card, Col } from "react-bootstrap";
-import UserListUI from '~/components/UserListUI.jsx';
+import UserListUI from '~/components/Enrollment/UserListUI.jsx';
   
 const ListLocationEnrollments = (user, enrollments, setDeletePath, setDeleteMessage, setDisplayConfirmationModal) => { return enrollments.map((enrollment) =>
   <Col key={enrollment.id}>
