+++ app/frontend/entrypoints/admin_enrollment_new.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import EnrollmentForm from '~/components/EnrollmentForm';
+import EnrollmentForm from '~/components/Enrollment/AdminOnly/EnrollmentForm';
 
 const container = document.getElementById('new');
 const root = ReactDOM.createRoot(container);
