+++ app/frontend/entrypoints/admin_student_new.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import NewStudentForm from '~/components/Student/NewStudentForm/NewStudentForm';
+import NewStudentForm from '~/components/Student/NewStudentForm';
 
 const container = document.getElementById('new');
 const root = ReactDOM.createRoot(container);
 
-root.render(<NewStudentForm admin="" />);
+root.render(<NewStudentForm admin="/admin" />);
