+++ app/frontend/entrypoints/confirmation.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import ConfirmationPage from '~/components/ConfirmatonPage/ConfirmationPage';
+import ConfirmationPage from '~/components/Confirmation/ConfirmationPage';
 
 const container = document.getElementById('confirmation');
 const root = ReactDOM.createRoot(container);
