+++ app/views/pages/home.html.erb
@@ -1,2 +1,3 @@
 <%= vite_javascript_tag 'home.jsx' %>
+
 <div id="home" />
