+++ vite.config.ts
@@ -4,5 +4,5 @@ import RubyPlugin from 'vite-plugin-ruby'
 export default defineConfig({
   plugins: [
     RubyPlugin()
-  ]
+  ],
 })
